v00_buffer_t

+++ src/v00/ut_lib_gtest.cpp
@@ -1,7 +1,8 @@
 #include <gtest/gtest.h>
 #include <thread>
-#include <neutrino_mock.hpp>
+//#include <neutrino_mock.hpp>
 
+#include <neutrino_transport.hpp>
 #include <neutrino_producer.hpp>
 
 using namespace neutrino::impl;
@@ -24,10 +25,10 @@ TEST(neutrino_nanoepoch, interval)
     std::this_thread::sleep_for(sleep_nanoseconds);
     auto x2 = neutrino_nanoepoch();
 
-    ASSERT_TRUE((x2 - x1) >= sleep_nanoseconds.count());
-    ASSERT_TRUE((x2 - x1) < sleep_nanoseconds_max.count());
+    ASSERT_TRUE((x2 - x1) >= (uint64_t)sleep_nanoseconds.count());
+    ASSERT_TRUE((x2 - x1) < (uint64_t)sleep_nanoseconds_max.count());
 }
-
+/*
 struct neutrino_general_workflow_tests : public ::testing::Test
 {
     const uint64_t checkpoint_id_1 = 1;
@@ -332,3 +333,4 @@ TEST_F(neutrino_general_workflow_tests, context_helper_exception_and_normal_inte
     validate_context_helper_exception_and_normal_interleaved<transport::frame_v00::known_encodings_t::BINARY_NETWORK>();
     //validate_context_helper_exception_and_normal_interleaved<transport::frame_v00::known_encodings_t::JSON>();
 }
+*/
